requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
aiosqlite==0.20.0
jinja2==3.1.4
python-dotenv==1.0.1
httpx==0.28.1
beautifulsoup4==4.12.3
pytrends==4.9.2
anthropic==0.42.0
python-multipart==0.0.18
